refactor: Encapsulate chapter read status calculation in a use case
This commit introduces the `CalculateAllChaptersReadStatusUseCase` to encapsulate the logic for determining the read status of all chapters for a given day.

Key changes include:
*   Creating the new `CalculateAllChaptersReadStatusUseCase` which centralizes the logic to check if a chapter or a range of verses is marked as read.
*   Moving the read status calculation logic out of the `PassageList` composable and into this dedicated use case.
*   Updating `DayUiStateFlowFactory` to use the new use case to pre-calculate the read status for all chapters and include it in the `DayUiState`.
*   Simplifying the `PassageList` composable, which now receives the calculated read status map directly from the UI state.
*   Registering the new use case in the dependency injection graph.
