+++ feature/day/src/commonMain/kotlin/com/quare/bibleplanner/feature/day/presentation/model/DayUiState.kt
@@ -12,5 +12,6 @@ internal sealed interface DayUiState {
         val books: List<BookDataModel>,
         val datePickerUiState: DatePickerUiState,
         val formattedReadDate: String?,
+        val chapterReadStatus: Map<Pair<Int, Int>, Boolean>,
     ) : DayUiState
 }
